Created new feature-StrStack branch to develop a StrStack facade for shunting algorithm.
Created two classes: TestStrStack and StrStack.

No need to repeat TDD process on this particular feature as it is virtually identical to
feature-NumStack. As such, I reused the code and of course, made the relevant changes such as
pushing string instead of floats on a StrStack instance.

I have asked about this and was told all three classes: NumStack, OpStack and StrStack are very similar.